retrived static contents list
