+++ src/components/Gallery/Gallery.js
@@ -1,14 +1,19 @@
 import React from 'react';
+
+import Aux from '../../hoc/Aux';
 import  './Gallery.css';
 
 const gallery = (props) => {
     return (
-        <div className="Gallery">
+        <Aux>
+            <div className="Gallery">
                 <p onClick={props.getTitle}>Title: {props.title}</p>
                 <p>Description: {props.description}</p>
                 <p>{props.children}</p>
                 <p>Rating: {props.rating} <input type="text" onChange={props.review} value={props.rating} /></p>
-        </div>
+            </div>
+        </Aux>
+        
     )
 };
